Speed up callsite_id generation
for loops only produce optimal code for object arrays, not for integer arrays.
Use nqp::iterator and shift_i instead where appropriate. Also consolidate the
loops a bit. It shouldn't matter where in the callsite id we write the index of
an argument name. We use it as an opaque id anyway and only to look up a hash
entry.