Fix compilation of conditionals not actually coercing condition to full width
The code created a coercion register but never actually wrote to it. I can only
assume that register re-use led to a useful value ending up in $coerce-reg by
accident. This is supported by the experimental allocation of a dummy register
which breaks NQP's build.

Unfortunately .coerce will release the source register unconditionally so we
cannot just use the condition's result_reg as We do need that later for
coercion to the conditional's result type.