Fix loss of some line number annotations
The annotations live primarily on QAST::Stmts and QAST::Stmt nodes.
However, if we had such a node directly as the chlid of a node that was
not itself such a node, we'd drop the annotation. This, in combination
with some tree simplification done by the Rakudo optimizer, meant that
very simple routines like:
    method foo($a) {
        $a ~ "foo"
    }
Could end up with their line of code being reported as the line of the
method declaration, not the line of code inside of the method. This led
to less accurate line number reporting, as well as oddities in test
coverage data and debug breakpoint setting.