Force latebound lexical lookups in dyncomp'd code
When we compile some code to run it at BEGIN time, and a closure is
taken and retained there, then it may have some connections to the
faked-up lexical environment it was compiled in. In such cases, force
lookups of lexicals to be late-bound, so that we do not get wrong
indexes. Fixes the issue exposed in
https://github.com/rakudo/rakudo/issues/2684.