Start explicitly marking out mixin types
MoarVM will start requiring types that are the target type of a rebless
(only used for mixins) to up-front declare they will be used that way.
This means it can assume anything that is not such a type - the vast
majority of cases - will never have extra data attached to them, which
can be used to generate simpler attribute access code.

While it's only MoarVM that will immediately use this, it's quite
possible that other backends will be able to make use of such things, so
it is done in a cross-backend way.

This also needed a rebootstrap of both the JVM and MoarVM backends.