[JVM] Add second version of bindlexdyn as bandaid
nqp::bindlexdyn expects the name as the first and the value as the
second argument (on MoarVM, too). It looks like that op wasn't used
in rakudo-j -- until https://github.com/rakudo/rakudo/commit/69c535ecfb

Since bindlexdyn is currently used in NQP's src/vm/jvm/QAST/Compiler.nqp
we need a change there and a rebootstrap in order to remove the old
version of bindlexdyn. For now, this should fix the Rakudo build.