[JVM] Let CONTROL do it's work in nqp::handle
Before, a labeled exception wasn't caught by CONTROL.

Again, I'm not very sure about this patch, but it helps with one
failing test and has no (additional) fallout for spectest.
Please revert if it's the wrong thing to do.

There might be other cases where unwind_check is called without
passing a value for $handler_cares. The check for $handler_cares
was introduced back in 2014 with 1783628e66.