Add method_names and method_order to NQPConcreteRoleHOW
This is used to iterate over the role's methods in order of declaration.
Though in NQP the same could be achieved through HOW.methods and asking the
methods for their name, in Perl 6 it's possible for the method's name to be
different from the way it's referred to by the class or role it's added to.
Adding these accessors keeps compatibility between NQP's and Perl 6's meta
object protocols.