[truffle] Implement (g|s)etcodename
However, currently getcodename only works if you've explicitly called
setcodename on the code ref first. Setting the name needs to be added to
the NQPCodeRef constructor (which is easy), but actually getting a real
value to put in there I haven't figured out how to do yet, and may
require implementing something like CodeRefAnnotation in/for truffle.