Consistently make the `.` in `<.[x]>` a no-op
In https://github.com/rakudo/rakudo/issues/3208 it was reported that we
got compilation errors from typing `<.[\d]>`, while the `.` was silently
ignored in `<.[a]>`. It turns out that the compiler would treat whatever
`.` was applied to as if it was going to be a subrule call, and when it
was something simple enough, we got away with that since the subtype was
ignored. However, when it was something like a cclass that wanted the
subtype, compilation would be broken.

The issue suggests making it an error. I'm sympathetic to that, however
doing so even breaks compilation of NQP and Rakudo - which do use things
like `<.[abc]>`. The `.` is actually redundant, so it's not entirely
clear why they do this. However, if making it an error breaks building
NQP/Rakudo, one would imagine the effects on the ecosystem will be no
better. Thus, we'll make it consistently ignored.