Save a bunch of allocations in MoarVM::StringHeap
All we did with the @!strings array was adding to it and collecting them into
the bytecode buffer at the end of compilation. Instead of a list of buffers, we
may as well collect them into one buffer that will later be written into the
mbc buffer in one go. This way we don't have to allocate for every single
string on the heap but can rely on the VM to grow the buffer in sensible steps.