Write header last instead of first and fixing up
The previous changes meant that the header had some incorrect values
when it was written (because the entire file wasn't being created in
memory before the values in the header were calculated), so we went back
and overwrote the incorrect ones. This changes it to write the header
at the end, so it has all the correct values available.