Improve error message for erroneous <.panic()> syntax
Fix for NQP issue GH #632

The issue was triggered by a parser error after a reformat of file
"src/HLL/sprintf.nqp" in preparation for sprintf changes. Wrapping
line 21 before the last '>' on the line fired an exception with the
following message:

    $ make
    +++ Preparing MOAR build directories
    ++++++ Building MOAR backend
    +++ Generating	gen/moar/stage1/NQPHLL.nqp
    +++ Compiling	gen/moar/stage1/NQPHLL.moarvm
    Unable to parse expression in metachar:sym<assert>; couldn't find final '>'  at line 3039, near "\n> ]\n     "

The message was very helpful to me until it was explained by Jonathan
that such whitespace is not legal in the panic token. Thus I filed the
issue and worked on a PR to improve the message. The message was
enhanced by changes in file:

+ src/HLL/Grammar.nqp

Using the message enhancement required generating a new NQP stage0 in
order to see its effect (see instructions in file
"docs/bootstrapping.pod"). After creating the same line wrap as before
the new message is seen to be:

    $ make
    +++ Preparing MOAR build directories
    ++++++ Building MOAR backend
    +++ Generating	gen/moar/stage1/NQPHLL.nqp
    +++ Compiling	gen/moar/stage1/NQPHLL.moarvm
    Unable to parse expression in metachar:sym<assert>; couldn't find final '>'
    '<.panic()>' cannot have whitespace outside '()' at line 3046, near "\n> ]\n
       at gen/moar/stage2/NQPHLL.nqp:1029  (src/vm/moar/stage0/NQPHLL.moarvm:panic)