Fix t/nqp/114-pod-panic with a relocated `nqp` executable
As `nqp::execname` is an op code not available on JVM and thus resulting
in a compile time error, a quite dirty workaround is used. We hide the op
in an eval and retrieve the value via `die`.

Also there is no sane way to retrieve the current nqp executable name
available on the JVM, so skip that test there.