MoarVM: Allow for explicitly marking the mainline_frame in the bytecode
Previously the convention was that frame 0 is always the mainline_frame (to be
distinguished from the main_frame). With frames created by nested compilation
(e.g. BEGIN time EVAL) added to the comp unit, this is no longer true. So add
a new field to the next bytecode version to mark that frame explicitly.