Add a nested compiler's load_dependency_tasks to the outer compiler's
If a nested compilation creates code objects that should make it into the
precompilation file, we also need to add its load_dependency_tasks, as those
are responsible for establishing the outer lexical scope (e.g. the setting).
Otherwise we will end up with errors like trying to call a VMNull.