[JVM] Make gather work with last/next/redo in loop
This helps with a long standing bug on the JVM backend:
https://github.com/rakudo/rakudo/issues/3634

If last/next/redo were called within an extra block (marked as
'immediate' or 'immediate_static'), that call was translated to
a direct JVM level call. In this case the produced UnwindException
never hit it's handler.

I'm not sure this is a good way to fix the problem. But at least
it points at the relevant code.

Another caveat: If last/next/redo is used with a label, we still
get an uncaught UnwindException.