Fix retrieving template.html location
Now that NQP and Rakudo are relocatable we must not rely on a build time
path. NQP home is meant to always point to the actual runtime path.
For this purpose introduce the home-dir() method on the Compiler object
which will determine the current NQP home directory at runtime. This is
analogous to a feature Rakudo also has (Rakudo home).

On JVM the necessary vm command line argument needed to enable the VM to
know the scripts executable path was not set. Do so now.