Register a plain HLL::Compiler object under the 'HLL' name
This allows the code in NQP to use functionality in the HLL Compiler.
Before that wasn't possible as it wasn't known which name the subclassed
Compiler object would have.

Use that generic name to access it from NQP code. In particular in the
logic that writes the `--profile` output.

This fixes rakudo/rakudo#3824