[JVM] Implement op 'execname'
This is a bit hacky, but it should enable us to use nqp::execname
in Rakudo's code base uniformly on all backends.

Currently we use nqp::jvmgetproperties(){nqp.execname} for NQP
and nqp::jvmgetproperties(){perl6.execname} for Rakudo.
Those properties are passed to the JVM via the runner scripts nqp-j
and rakudo-j, respectively.

I've also added support for raku.execname, so perl6.execname could
be easily removed in the future.