Serialization format 23 for nqp-JS
nqp-JS is currently a cross-compiler, and so needs to be able to read the
serialization output of MoarVM. As of Feb 2020, commit 9709537d90d61529:
    Serialize the names of closures

    The name is held per code object, meaning different closures may carry a
    different name. This is notably true of auto-generated `proto` methods.
    However, we did not preserve this during serialization, meaning that
    Raku language users could see things like AUTOGEN showing up in both
    backtraces and introspection. Fix this by serializing the name along
    with the closure. This, one a spectest is added, will let us resolve
    https://github.com/rakudo/rakudo/issues/3323.

the MoarVM serialization format was bumped from version 22 to version 23,
as it now also serializes closure names.

This broke the nqp-JS build, but it seems that no-one noticed.

This fixes the build at 2020-07 (the current release), but unfortunately
additional fixes are needed for changes now made on master.