[JVM] Avoid decoding errors for multibyte chars
The old implementation tried to decode ByteBuffer after ByteBuffer.
If the bytes of a multibyte character happened to be splitted
between two ByteBuffers, the decoding failed (for no good reason).

Please note that we are still in trouble if the remaining bytes
of a multibyte char have not arrived yet. But this seems to be a
step in the right direction.