[JVM] Remove superfluous check for incomplete char
It looks like the real problem was that we did signal endOfInput
to CharsetDecoder.decode when using nqp::takeavailablechars.
After the fix in 64479933f9 we don't need the whole machinery
for testing for incomplete characters -- CharsetDecoder.decode
already does that check and does only report an error if
endOfInput is set.