[JVM] Discern preliminary vs final decoding error
If CharsetDecoder.decode() is not able to fully decode the current
input buffer and there is more input available, the reason could be
that the current input buffer ends with bytes, that *could* be part
of a multibyte character. In order to make progress, the remaining
bytes are merged with the bytes of the next input buffer and the
decoding process is continued.