Sort all entries alphabetically for easier lookup
Also add a "sort-ops" script that will make sure that all sections
and all ops in "ops.markdown" are in alphabetical order, except
for the "Miscellaneous sections" that will always be listed last.

Add a new section is easy: add the section at the end, move any
ops that you want in there in that section, and run the "sort-ops"
script again.  It will automatically add the new section to the
table of contents at the beginning.