Fix parallel compilation occasionally losing frames
CUIDs must as their name suggest be unique within a computation unit. We rely
on this when we use CUIDs as keys for lookup hashes for the static frames into
which we compile blocks.

When multiple compilations (like EVAL) are run in parallel, it was possible
that we lost an increment of the global $cur_cuid which could cause us to
assign the same CUID to 2 different blocks. This would further lead us to
losing on of the associated frames and causing the other one to be referenced
instead.

Fix by protecting the counter by a lock. This could be optimized by using
atomic operations instead.