Remove long obsolete code
8 years ago in commit b6c88ffae3de22ead132d5751ab11517ebacd4dc the %caches
lexical was turned into an attribute. Since now every instance of the meta
class has its own caches hash, there was actually no longer a need to retain
the nested hash keyed on the meta object's memory address. The cache method
was adapted accordingly, but the obsolete initialization in add_method was
kept needlessly. Lay it to rest, finally.