Eliminate the ops sec_n, asec_n and sech_h, which are no longer used.
These hyperbolic functions are not provided directly by libc, so have to be
implemented in terms of functions which are. There's no mathematical
difference in expressing the same identities in the Rakudo setting (it's
going to dispatched as the same IEEE operations) but it's better to do it
once there instead of once per backend.

Rakudo stopped using these ops in the 2021.02 release; now now remove them
from NQP and MoarVM.

The JS implementation of sech this commit removes is curious - it's
different from the JVM implementation, MoarVM's C implementation, all the
other JS implementations I can find online, and the "textbook" definition:

    sech 𝑥 = 1 / cosh 𝑥

The relevant commits don't give any clues.

However, I *believe* it to be mathematically equivalent - after a lot of
head scratching and dead ends, the identities that one needs to re-arrange
from it to the canonical definition above are that cosh 2𝑥 = cosh²𝑥 + sinh²𝑥
and cosh²𝑥 = 1 + sinh²𝑥. Hence *why* - it makes two cosh calls whereas the
"obvious" formula uses only one, *and* it has to special case ±∞. I *think*
that it might produce "better" results (ie - not just 0.0) for inputs whose
output /could/ be correctly expressed as an IEEE subnormal. But that's the
only reason I can see for doing this; I can't find it described anywhere,
and it seems a lot of work for a corner case of a corner case, which isn't
even going to be mathematically rigorous. And the fact that it needs to
special case Inf to avoid evaluating 0.0/0.0 and hence returning NaN makes
me wonder if it (still) produces similar results for some very large inputs.

Really, if you care about extreme results, you should be calculating in
symbolic algebra anyway...

Of course, if I'm wrong on this, then this approach ought to be used in the
Rakudo setting (along with the explanation of why it's objectively better.)