[JVM] Make nqp::unbox_i on NativeCall object work
It is supposed to report whether the library was loaded or not. This
is a port of https://github.com/MoarVM/MoarVM/commit/1b0950e3e9 and
is needed since https://github.com/rakudo/rakudo/commit/9ef7a7e9dd.

nine++ gave the hint for the concrete implementation on the JVM.

Please note, that the usage of nqp::unbox_i($!call) in
lib/NativeCall.rakudmod don't work yet. They result in a
NullPointerException -- but that's probably a different bug. At least
the following code works now:

  use nqp;
  use NativeCall;
  my class native_callsite is repr('NativeCall') { };
  my $bar = native_callsite.new;
  say nqp::unbox_i($bar);  ## 0