[JVM] Keep correct REPR.name for native arrays
It turned out that f3dc977 doesn't really fix the usage of native
arrays in the core setting. For some reason during the compilation of
blib/CORE.d.setting.jar native array are deserialized twice.

The first time worked as expected, but since we did reset REPR.name
to VMArray, the second time went wrong just as it used to fail before
f3dc977.

Since REPR.name is also used for nqp::reprname, it seems cleaner to
leave it untouched and add a different field (vmarray_name) to REPR
instead.