Remove the --rxtrace feature
It would need implementing somewhat differently under new-disp, but also
seems very rarely used (the same kind of information is available by
looking at --target=parse). Furthermore, the way it was done incurred
costs even when not in use on one very hot path (`find_method`, which
got hotter on new-disp) and also on every language switch (also fairly
hot, given it happens on each string literal or regex we encounter
when parsing), which is quite a lot to pay for something barely used.
Drop it; if somebody really misses it, we'll find out and can find a
better way to do it.