Avoid lots of dispatch program creation at startup
The QAST operations setup previously made a method call for every single
op. This resulted in producing a dispatch program that was only used
once ever. For most simple mappings, we can instead use a constant table
of them and add them in a loop. This knocks a few percent off startup
time, plus avoids the wasted memory of a bunch of single-use dispatch
programs and callsites.