Cache flat method tables on meta-objects
We produce these in the situation that a dispatch is going megamorphic
and we want to handle that by doing a hash lookup on method names. In
the case that there are also loads of types, we currently fill up the
callsite cache with an entry per encountered type (this needs to be
addressed). Once it fills, we'd end up recalculating this table over and
over again. Further, even in the non-full cache case, we could calculate
and install it at many sites, which is a waste of time and memory. Thus,
calculate it once and share it among the various caches.