Fix source-version not picking up changes to MAST::NodesOps
The source-version is used to determine whether rakudo needs to rebuild and for
fetching the correct version of precomp files. It must change with any changes
to the built binaries (including bytecode files). MAST::Nodes and MAST::Ops are
not part of NQP's sources but supplied by MoarVM instead, but they still factor
into the build result and must be included in source-version.