Fix Missing or wrong version of dependency 'gen/moar/stage2/NQPHLL.nqp' errors
Complaints about gen/moar/stage2/NQPHLL.nqp being the wrong version can only
happen if that file is not included in the NQP source digest that is included in
Rakudo's compiler id. And in fact gen/moar/stage2/NQPHLL.nqp isn't included, at
least not directly. That file is generated by concatenating a number of source
files. So everything should be fine as long as all its sources are included in
the source digest.

However among a number of .nqp files from src/ (which are all included in the
digest), there's also gen/moar/stage2/nqp-config.nqp which is _not_ included in
the digest. It can't be, because that's where tools/build/gen-version.pl writes
the digest to.

Now there are few reasons for why that file would change. Obviously a change in
the src/ files would yield a different digest, thus changing nqp-config.nqp. But
since the digest already changed, we're still safe. However, there is also the
NQP version and a few paths. So any version change will change this file and
only this file. In turn this will change gen/moar/stage2/NQPHLL.nqp, thus
NQPHLL.moarvm which will ultimately lead to the dependency error.

Fix by including all information that's used to generate nqp-version.nqp in the
source digest.