Introduce Counter, a bounded native atomic integer
The metamodel needs a reasonably fast thread-safe counter. Atomic ops
are also easy to screw up, on the other hand. This adds a `Counter`
class to `NQPCORE.setting` to take care of this for ourselves, which
performs comparably to the semaphore that it was originally intended to
be.

This probably doesn't work on backends lacking native atomic integer
support. *Surely* we have a version of the JVM and node.js that makes
this possible by now?? Right???