[JVM] Try for a lazier LibraryLoader
We've been depending largely on byte *arrays* of entire files so far.
Now we prefer to mmap in small chunks when access to a file is given,
doing a heap allocation given raw byte input instead, waiting until the
last minute to dump the entire thing in memory in either case.

We introduce a ByteBufferedInputStream to help avoid repetitively
reading files into (potentially huge!) chunks of memory on its own. Its
`copy` method should *not* exist, but this class is a whole tangent on
its own, so just do the bare minimum of an implementation of an input
stream given a byte buffer.