[JVM] Replace IgnoreNameClassLoader with a parallel SerialClassLoader
We have a shiny new serialized XZ type of file to read, so we have no
worries as to versioning of how its read; let's exploit that. Java
enjoys having names for its defined classes, so put the name of the
class in the XZ's JAR entry comment field and use it when redefining it
on read. We wind up with null if none can be found anyways, giving us
prior behaviour for uncompressed serializations too.

We have a loadClass-ish method that doesn't follow Java's spec per se.
We have the choice of linking with resolveClass after building one with
a findClass-ish method; do so.

Slap a synchronized modifier on our new-ish loadByteClass method and now
there's nothing preventing us from having parallel class loading. Mark
our new SerialClassLoader as being parallel-capable and set
-XX:+AllowParallelDefineClass now that we have a somewhat more modern
JDK and now that there shouldn't be anything else preventing us from
doing so.