[JVM] Never load bulk files into memory from LibraryLoader.load
We *should be* able to InputStream our way out of keeping files in
memory, with the exception of when other parts of nqp do so beforehand,
thus the existence of the MemoryClassLoader. The SerialClassLoader wraps
an input stream to read from, which JarFileClassLoader extends with the
preliminary file loading. A similar FileClassLoader exists, which only
cares about decompressed classfiles.

Regardless, Rakudo builds are broken because xz's XZInputStream has its
bugs. Womp womp.