[JVM] Drop xz for lz4
xz now breaks Rakudo builds by its minimal, quirky InputStream
implementation in the form of XZInputStream, which doesn't play nice
with the default readAllBytes implementation, which we kinda need in
some form.

lz4, in comparison to xz and zlib, is more inefficient in terms of
compressed output size (NQPCORE.setting.jar is now 57KB compared to 52KB
w/ ZIP DEFLATE), but is nevertheless more efficient to compress and
decompress.  Because precomp files are apt to exist locally, hopefully
this tradeoff is worth it...