[JVM] Unify the disparate class caches in ByteClassLoader as a parent
ByteClassLoader tracks which classes are defined, while GlobalContext
tracks which are loaded. The LibraryLoader depends on the former, but
what if the latter already defined a class to be loaded (circling back
to itself via require, perhaps)? ByteClassLoader can keep track of all
classes loaded or defined by ourselves if it acts as a parent for
LibraryLoader's various classloaders, eliminating such a duplicate
definition. If we give StreamClassLoader and MemoryClassLoader a new
SerialClassLoader parent that takes care of class caching, we get to
dedup some logic there as well.