Add --gen-moar/--with-moar to build command
As pointed out in https://github.com/Raku/nqp/issues/792 the previously
advertised command

  $ perl Configure.pl --backends=moar,jvm

would just fail.

I've also changed the description of "--with-moar" further down, to
make it match the current behaviour: Giving a directory for that option
leads to an (failing) attempt to execute that directory with --version.
This could be a bug. Specifying the full path to the moar executable
works as expected.