Mark t/nqp/063-slurp.t as self-referential
Also, there is no need to accept two numbers as the number of chars
anymore. With 0412c22303 the JVM backend learned to count a CRLF
as one char as well. So even if the test file would be written with
CRLF as line endings, nqp::chars should give back the same number
as for LF as line endings.

This should fix https://github.com/Raku/nqp/issues/512.