Close handle before unlinking
On Windows a file can't be deleted if it is "busy", i.e. if there is an open handle to it. This start failing locally while making some IO changes in MoarVM, although I'm not sure why it hasn't always been failing.

This updates a file deletion test to close the file it is going to unlink before actually trying to unlink it.