Nano-optimize while loops
Instead of starting at -1 and always incrementing inside the check
(an optimization that turned out to be very useful in Raku), it
appears that this way ensure that there is 1 increment more than
really necessary.  And if there is no actual iteration possible,
this is relatively a lot.

Revert to starting at 0: this means no actual initialization of
the counter is needed (although at the VM level this still needs
to be done, of course) and move the increment to the end of the
loop.

This nibbles about .05 seconds of CPU time in make test.  Every
little bit helps.

Also remove some dead code.