Only add unique types to the typecache
One of the changes in 3291cb41ce862163cb18 was that recursion on
roles in the typecache was unlimited, rather than only one level
deep.  This tripped something in the Rakudo build causing it to
infiniloop on the build, presumably because of a type having a
role having that same type, etc. etc.

This commit adds a "seen" check for types before they're added to
what will make the typecache for the meta object.  Oddly enough,
the added check appears to have a positive affect on build times
of both NQP as well as rakudo.