Generalize handling of TWEAK
Adds a $!tweaks attribute to NQPClassHOW: a list with codeblocks to
be executed at the end of the local BUILDPLAN.  This should allow
TWEAK blocks from roles to be added to a class, and thus provide
a consistent means to initialize attributes of a role that is
consumed by a class.

Also add some basic testing of BUILD / TWEAK: they did not appear
to be tested as such yet.