Make the positional on "method_table" optional
And change all of the calls to "method_table" to not pass any
positional argument.

This reduces the footprint of these calls.  This is possible
because the positional argument was only there to allow for
.^method calls to work from Raku (as far as I can see).

And since this method (as many others) only looks at attributes
of the invocant, this will not break any code.