Make the value generator only run if really needed
Since the value generator can be relatively expensive, make sure
it only runs if it is really necessary: namely when the key is
really not in the cache after having unique access to it.

Also, the value generator may have race conditions of its own,
so it feels better to have this run inside the protected block.