Only recompile NQPP5QRegex if source files are modified
Because the NQPP5QRegex target generates the file it compiles, its
modification time is always new. Because the test target requires the
NQPP5QRegex target, it gets recompiled *every* time you run a test.
Adjust the modification time of the generated file back to the most
recent modification time of its source files, so it only gets recompiled
if one of them has changed.