Add "code_of_method" and "declares_method" methods
to NQPClassHOW.

- declares_method returns 1 if there's a method by that name in
  the method table (effectively wrapping an nqp::existskey)
- code_of_method returns the code object of the method by the given
  name, or nqp::null if there's no method by that name (effectively
  wrapping an nqp::atkey.

This is mostly to provide compatibility with methods by the same
name in Raku's ClassHOW, where the check is a little more involved
thanks to the existence of submethods.  By abstracting the lookup
it will now also be possible to remove a lot of checks for submethod
tables, as they can now be abstracted in methods with these names.