[JVM] Fix capturenamedshash to actually include objects
The condition `(flagged & CallSiteDescriptor.ARG_OBJ) != 0` will
never be true because ARG_OBJ is defined as 0.

It's not clear when this broke, and it also looks like capturenamedshash
is never really called anywhere -- at least my angry debug statements
never seemed to be output during the building of CORE.setting.