More tweaking fetching MRO
Alas, c92ed85ec11e2 didn't prevent the REA from crashing every now
and then.  Basically revert it and add a little bit of debugging
code that will sleep .1 second and then try again: this on the
assumption that it is somehow a race condition that makes the
$!mro attribute being nqp::null momentarily.

If this code still fires, I guess we can rule out a simple race
condition, but have to assume something more nefarious.