Add is-interactive to HLL::Compiler
This commit nukes the hard-to-grok distinction of
`interactive` vs `non-interactive` repl-modes and
replaces it with the following options:

    - `tty`
    - `process`
    - `disabled`

`disabled` is akin to what happens when you run `perl` or `python`,
it just sits there waiting for STDIN to complete with an <EOF>, then
runs the code. It's identical to what `non-interactive` means at the
moment.

`tty` is the default and is mostly presented to the user for
completeness' sake but it can be useful in a situation where the
user wants to interpreter to refuse to start outside of a valid TTY.

`process` is the same as `interactive` except that it explicitly
refers to the expectation that STDIN and STDOUT will *not* be
TTY. (The intention of `interactive` is to "bypass bypass
TTY detection").

These changes resolve the ambiguity that occurs when trying to
imagine a "non-interactive" REPL (an obvious oxymoron) or to
distinguish between a REPL that is started in a Proc without any
expectations of a TTY and a REPL that requires a TTY.